Extract parser expression grammar
